ml_service/requirements.txt
fastapi
uvicorn
torch
torchvision
timm
pillow
numpy
pandas
scikit-learn
python-multipart
opencv-python
